Fixes for comparisons between signed and unsigned values with new jb-mr2 compiler (for x86, ARM previously fixed in d74e41b1cce5373aa24fd2fbea735173f6113d5a)

+++ src/oat/runtime/x86/context_x86.h
@@ -43,7 +43,7 @@ class X86Context : public Context {
   }
 
   virtual uintptr_t GetGPR(uint32_t reg) {
-    CHECK_LT(reg, kNumberOfCpuRegisters);
+    CHECK_LT(reg, static_cast<uint32_t>(kNumberOfCpuRegisters));
     return *gprs_[reg];
   }
